+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement system info endpoint [#6]
 - implement maintenance mode endpoint [#18]
 - Implement export mode endpoint for multinode [#14] and classic mode [#20]
+- Implement configuration endpoint for classic ces [#16]
 
 ## Added
 - New Make-Target (`make debian`) to build a debian package to install the exporter in a classic CES [#12]